refactor(rte): extract llm runtime seam
